+++ src/ErrorPropagation/statistics.jl
@@ -7,7 +7,7 @@
 # standard error estimate:
 # Take the highest lvl with at least 32 bins.
 # (Chose 32 based on https://doi.org/10.1119/1.3247985)
-function _reliable_level(B::ErrorPropagator{T,N})::Int64 where {T,N}
+function _reliable_level(B::ErrorPropagator{T,N})::Int where {T,N}
     isempty(B) && (return 1)                # results in NaN in std_error
     i = findlast(x -> x >= 32, B.count)
     return something(i, 1)
